% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ethanol}
\alias{ethanol}
\title{Ethanol levels}
\format{
A data frame with 504 rows and 3 variables:
\describe{
\item{id}{unique id of the time series}
\item{class}{four classes: e35, e38, e40, e45}
\item{ts}{list column, containing a time series of length 1751 for each row}
...
}
}
\source{
\url{http://www.timeseriesclassification.com/description.php?Dataset=EthanolLevel}
}
\usage{
ethanol
}
\description{
Time series of ethanol with varying levels of alcohol. Four classes.
}
\keyword{datasets}
