% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prices}
\alias{prices}
\title{Share Price Increase}
\format{
A data frame with 53940 rows and 10 variables:
\describe{
\item{id}{unique id of the time series}
\item{class}{whether or not the time series increased or not}
\item{ts}{list column, containing a time series of length 60 for each row}
...
}
}
\source{
\url{http://www.timeseriesclassification.com/description.php?Dataset=SharePriceIncrease}
}
\usage{
prices
}
\description{
A dataset containing the prices and other attributes of almost 54,000
diamonds.
}
\keyword{datasets}
