% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstruct.R
\name{reconstruct}
\alias{reconstruct}
\title{Reconstruct time series}
\usage{
reconstruct(x, name, step, ...)
}
\arguments{
\item{x}{data frame}

\item{name}{name of the time series to reconstruct (the original column name
before \code{bake}).}

\item{step}{the trained step from the recipe}

\item{...}{tidy-selection of columns to reconstruct from}
}
\description{
Reconstruct the time series \code{name} using the trained \link{step_dct}
with the columns specified in \code{...}.
}
\details{
May be a lossy reconstruct of the time series.
}
