% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtw.R
\name{step_dtw}
\alias{step_dtw}
\title{Clustering time series using dynamic time warping}
\usage{
step_dtw(
  recipe,
  ...,
  role = "predictor",
  trained = FALSE,
  k = 4,
  preserve = FALSE,
  dtwclust = NULL,
  options = list(),
  skip = FALSE,
  id = recipes::rand_id("dtw")
)
}
\arguments{
\item{recipe}{A recipe object. The step will be added to the
sequence of operations for this recipe.}

\item{...}{One or more selector functions to choose which
variables are affected by the step. See \code{\link[recipes:selections]{selections()}}
for more details. For the \code{tidy} method, these are not
currently used.}

\item{role}{For model terms created by this step, what analysis
role should they be assigned?. By default, the function assumes
that the new columns created from the original variables will be
used as predictors in a model.}

\item{trained}{A logical to indicate if the quantities for
preprocessing have been estimated.}

\item{k}{The number of clusters, tunable.}

\item{dtwclust}{A list of \code{\link[dtwclust:TSClusters-class]{TClusters}}
objects for each list of time series passed to the step, created once
the step has been trained.}

\item{options}{A list of options for \code{\link[splines:bs]{splines::bs()}}
which should not include \code{x}, \code{degree}, or \code{df}.}

\item{skip}{A logical. Should the step be skipped when the
recipe is baked by \code{\link[recipes:bake.recipe]{bake.recipe()}}? While all operations are baked
when \code{\link[recipes:prep.recipe]{prep.recipe()}} is run, some operations may not be able to be
conducted on new data (e.g. processing the outcome variable(s)).
Care should be taken when using \code{skip = TRUE} as it may affect
the computations for subsequent operations}

\item{id}{A character string that is unique to this step to identify it.}
}
\description{
Clusters a time series using \code{\link[dtwclust:tsclust]{dtwclust::tsclust}}.
The number of clusters is tunable with \code{k}. Additional parameters
can be set using \code{options}.

Adds a single column to new data, with integers 1-k identifying the cluster.
}
