% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fft.R
\name{step_fft}
\alias{step_fft}
\title{Extract discrete Fourier transform coefficients}
\usage{
step_fft(
  recipe,
  ...,
  role = "predictor",
  trained = FALSE,
  k = 4,
  series = NULL,
  skip = FALSE,
  id = recipes::rand_id("tff")
)
}
\arguments{
\item{recipe}{A recipe object. The step will be added to the
sequence of operations for this recipe.}

\item{...}{One or more selector functions to choose variables
for this step. See \code{\link[recipes:selections]{selections()}} for more details.}

\item{role}{For model terms created by this step, what analysis role should
they be assigned? By default, the new columns created by this step from
the original variables will be used as \emph{predictors} in a model.}

\item{trained}{A logical to indicate if the quantities for
preprocessing have been estimated.}

\item{k}{The number of discrete Fourier transform coefficients}

\item{series}{Name of column to transform,  created once the step has been trained.}

\item{skip}{A logical. Should the step be skipped when the
recipe is baked by \code{\link[recipes:bake]{bake()}}? While all operations are baked
when \code{\link[recipes:prep]{prep()}} is run, some operations may not be able to be
conducted on new data (e.g. processing the outcome variable(s)).
Care should be taken when using \code{skip = TRUE} as it may affect
the computations for subsequent operations.}

\item{id}{A character string that is unique to this step to identify it.}
}
\description{
In almost all cases, use \link{step_dct} instead.
The discrete cosine transform has better compaction in general.
DCT coefficients are always real, while DFT returns complex numbers,
which increases the number of dimensions.
}
\references{
Sayood, K. Introduction to Data Compression
}
