% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prices}
\alias{prices}
\title{Share Price Increase}
\format{
A data frame with 965 rows and 3 variables:
\describe{
\item{id}{unique id of the time series}
\item{class}{whether or not the time series increased or not}
\item{ts}{list column, containing a time series of length 60 for each row}
...
}
}
\source{
\url{http://www.timeseriesclassification.com/description.php?Dataset=SharePriceIncrease}
}
\usage{
prices
}
\description{
Time series of price increases over a 60 day period.
}
\keyword{datasets}
